pyproject.toml
[tool.poetry]
name = "muravleva-bookshelf-django"
version = "0.1.0"
description = ""
authors = ["Muravleva"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3"
django-cors-headers = "3.7.0"
djangorestframework = "3.12.2"
whitenoise = "5.2.0"
dj-database-url = "0.5.0"
djoser = "2.1.0"
drf-yasg = "1.20.0"
psycopg2-binary = "2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
